created fetchUserAdmin
