+++ client/src/providers/AuthProvider.js
@@ -32,8 +32,10 @@ export class AuthProvider extends React.Component {
   handleUserLogin = (user, history) => {
     Axios.post("/api/auth/sign_in", user)
       .then((res) => {
+        this.handleAdminLogout()
         this.setState({ user: res.data.data });
         console.log(this.state.user);
+        localStorage.setItem('token-type','user')
         history.push("/");
       })
       .catch((err) => {
@@ -46,8 +48,10 @@ export class AuthProvider extends React.Component {
     debugger;
     Axios.post("/api/admin_auth/sign_in", admin)
       .then((res) => {
+        this.handleUserLogout()
         this.setState({ admin: res.data.data });
         console.log(this.state.admin);
+        localStorage.setItem('token-type','admin')
         history.push("/");
       })
       .catch((err) => {
@@ -60,6 +64,7 @@ export class AuthProvider extends React.Component {
     Axios.delete("/api/auth/sign_out")
       .then((res) => {
         this.setState({ user: null });
+        localStorage.removeItem('token-type')
         history.push("/login");
       })
       .catch((err) => {
@@ -71,6 +76,7 @@ export class AuthProvider extends React.Component {
     Axios.delete("/api/admin_auth/sign_out")
       .then((res) => {
         this.setState({ admin: null });
+        localStorage.removeItem('token-type')
         history.push("/admin-login");
       })
       .catch((err) => {
